Refactored code

+++ src/services/issue.js
@@ -1,28 +1,28 @@
+import mongoose from "mongoose";
 import { Issue } from "../models/issue.js";
 
 // Create an issue
-export const createIssue = (issue) => {
-    return Issue.create(issue);
+export const createIssue = async (issue) => {
+    return await Issue.create(issue);
 }
 
 // Get all issues create by the specific user
-export const getIssuesCreatedByUser = (userId) => {
-    return getIssues({ createdBy: userId });
+export const getIssuesCreatedByUser = async (userId) => {
+    return await getIssues({ createdBy: userId });
 }
 
 // Get all issues based on the filter input
-export const getIssues = (filter) => {
-    return Issue.find(filter);
+export const getIssues = async (filter) => {
+    return await Issue.find(filter);
 }
 
 // Update the issue
 export const updateIssue = async (id, issue) => {
     // Get existing issue
-    const existingIssues = await Issue.find({ _id: id });
-    if (existingIssues.length === 0) {
-        throw new Error('Issue not found');
+    const existingIssue = await getIssueById(id);
+    if (!existingIssue) {
+        throw new NotFoundError(`Issue not found with id: ${id}`);
     }
-    const existingIssue = existingIssues[0];
     // Update the issue
     for (let key in issue) {
         existingIssue[key] = issue[key];
@@ -33,7 +33,10 @@ export const updateIssue = async (id, issue) => {
 
 // Delete the issue
 export const deleteIssue = async (id) => {
-    const issue = await Issue.findById(id);
+    const issue = await getIssueById(id);
+    if (!issue) {
+        throw new NotFoundError(`Issue not found with id: ${id}`);
+    }
     await Issue.deleteOne({ _id: id });
     return issue;
 }
@@ -41,4 +44,11 @@ export const deleteIssue = async (id) => {
 // Get issue by id
 export const getIssueById = async (id) => {
     return await Issue.findById(id);
+}
+
+class NotFoundError extends Error {
+    constructor(message) {
+        super(message);
+        this.name = "NotFoundError";
+    }
 }
